@@ -22,18 +22,18 @@ package_dir =
     =src
 packages = find:
 zip_safe = True
-python_requires= >=3.10
+python_requires= >=3.11,<3.14
 install_requires =
     astraeus@git+https://github.com/kevin218/Astraeus@main
     astropy
     astroquery
     batman-package
-    bokeh<3.0
-    catwoman@git+https://github.com/taylorbell57/catwoman@master
+    bokeh
+    catwoman@git+https://github.com/KathrynJones1/catwoman@master
     ccdproc
     celerite2 # Needed for GP
     corner
-    crds<12 # Upper limit needed to avoid bugs with crds.get_context_name
+    crds
     dynesty>1.0 # Lower limit needed for specific arguments
     emcee>3.0.0 # Lower limit needed for specific arguments
     exotic-ld==3.2.0 # Lower limit needed for updated JWST sensitivity files, upper limit needed for breaking changes
@@ -42,14 +42,17 @@ install_requires =
     h5py
     lmfit
     matplotlib>=3.6 # Lower limit needed for set_layout_engine()
-    mc3 # Needed for uncertainties in the Allan variance plots in S5
-    numpy>=1.20.0,<1.25 # Upper limit needed for Apple silicon
+    mc3 # Needed for uncertainties in the RMS time-averaging plots in S5
+    numpy>=1.20.0
     pandas
-    pastasoss>=1.2.1
+    pastasoss>=1.2.2 # Lower limit needed to allow for support with pymc3
     photutils
+    # planet-harmonica # Currently only supports for python 3.8 -- 3.10
+    planet-harmonica@git+https://github.com/DavoGrant/harmonica@main
     requests
-    scipy>=1.8.0 # Lower limit needed for scipy.spatial.QhullError
+    scipy>=1.12.0 # Lower limit needed for stcal.ramp_fitting.likely_fit
     setuptools_scm
+    shapely
     stdatamodels
     svo_filters
     tqdm
@@ -61,7 +64,7 @@ where = src
 
 [options.extras_require]
 jwst =
-    jwst==1.15.1
+    jwst==1.18.0
     stcal>=1.0.0 # Lower limit needed for our create_integration_model function
 hst =
     image_registration@git+https://github.com/keflavich/image_registration@master # Need GitHub version to avoid np.float issue
@@ -93,6 +96,7 @@ pymc3 =
     xarray<2023.10.0 # Upper limit needed for numpy<1.22
 jupyter =
     ipykernel
+    ipympl # Needed for the interactive matplotlib backend
     ipython
     jupyter
 
