@@ -59,7 +59,9 @@ def rampfitJWST(eventlabel, ecf_path=None, input_meta=None):
     # First apply any instrument-specific defaults
     if meta.inst == 'miri':
         meta.set_MIRI_defaults()
-    elif meta.inst in ['nircam', 'nirspec', 'niriss']:
+    elif meta.inst == 'niriss':
+        meta.set_NIRISS_defaults()
+    elif meta.inst in ['nircam', 'nirspec']:
         meta.set_NIR_defaults()
     # Then apply instrument-agnostic defaults
     meta.set_defaults()
@@ -185,6 +187,7 @@ class EurekaS1Pipeline(Detector1Pipeline):
         self.jump.maximum_cores = meta.maximum_cores
         self.jump.rejection_threshold = meta.jump_rejection_threshold
         self.jump.minimum_sigclip_groups = meta.minimum_sigclip_groups
+        self.clean_flicker_noise.skip = meta.skip_clean_flicker_noise
         self.gain_scale.skip = meta.skip_gain_scale
 
         # Instrument Specific Steps
@@ -204,17 +207,21 @@ class EurekaS1Pipeline(Detector1Pipeline):
             self.reset.skip = meta.skip_reset
             self.rscd.skip = meta.skip_rscd
             self.emicorr.skip = meta.skip_emicorr
+            self.emicorr.algorithm = meta.emicorr_algorithm
 
         # Define ramp fitting procedure
         self.ramp_fit = Eureka_RampFitStep()
         self.ramp_fit.algorithm = meta.ramp_fit_algorithm
         self.ramp_fit.maximum_cores = meta.maximum_cores
         self.ramp_fit.skip = meta.skip_ramp_fitting
+        self.ramp_fit.firstgroup = meta.ramp_fit_firstgroup
+        self.ramp_fit.lastgroup = meta.ramp_fit_lastgroup
+        self.ramp_fit.suppress_one_group = meta.ramp_fit_suppress_one_group
         self.ramp_fit.s1_meta = meta
         self.ramp_fit.s1_log = log
 
         # Default ramp fitting settings
-        if self.ramp_fit.algorithm == 'default':
+        if self.ramp_fit.algorithm in ['OLS', 'OLS_C']:
             self.ramp_fit.weighting = meta.default_ramp_fit_weighting
             # Some weighting methods need additional parameters
             if self.ramp_fit.weighting == 'fixed':
