@@ -78,8 +78,8 @@ def read(filename, data, meta, log):
     elif len(int_times['int_mid_BJD_TDB']) == 0:
         # There is no time information in the simulated NIRSpec data
         if meta.firstFile:
-            log.writelog('  WARNING: The timestamps for simulated NIRSpec data '
-                         'are not in the .fits files, so using integration '
+            log.writelog('  WARNING: The timestamps for simulated NIRSpec data'
+                         ' are not in the .fits files, so using integration '
                          'number as the time value instead.')
         time = np.linspace(data.mhdr['EXPSTART'], data.mhdr['EXPEND'],
                            data.intend)
