@@ -6,7 +6,7 @@ from matplotlib import rcParams
 try:
     from mc3.stats import time_avg
 except ModuleNotFoundError:
-    print("Could not import MC3. No Allan variance plots will be produced.")
+    print("Could not import MC3. No RMS time-averaging plots will be made.")
 import corner
 from scipy import stats
 import fleck
@@ -17,6 +17,11 @@ try:
     import starry
 except ModuleNotFoundError:
     # PyMC3 hasn't been installed
+    pass
+try:
+    from harmonica import HarmonicaTransit
+except ModuleNotFoundError:
+    # Harmonica hasn't been installed
     pass
 import warnings
 warnings.filterwarnings("ignore", message='Ignoring specified arguments in '
@@ -42,7 +47,6 @@ def plot_fit(lc, model, meta, fitter, isTitle=True):
         The name of the fitter (for plot filename).
     isTitle : bool; optional
         Should figure have a title. Defaults to True.
-
     """
     if not isinstance(fitter, str):
         raise ValueError(f'Expected type str for fitter, instead received a '
@@ -163,13 +167,6 @@ def plot_phase_variations(lc, model, meta, fitter, isTitle=True):
         The name of the fitter (for plot filename).
     isTitle : bool; optional
         Should figure have a title. Defaults to True.
-
-    Notes
-    -----
-    History:
-
-    - September 12, 2022 Taylor Bell
-        Initial version.
     """
     if not isinstance(fitter, str):
         raise ValueError(f'Expected type str for fitter, instead received a '
@@ -309,7 +306,7 @@ def plot_phase_variations(lc, model, meta, fitter, isTitle=True):
 
 
 def plot_rms(lc, model, meta, fitter):
-    """Create an Allan variance plot to look for red noise. (Figs 5301)
+    """Create a RMS time-averaging plot to look for red noise. (Figs 5301)
 
     Parameters
     ----------
@@ -321,15 +318,6 @@ def plot_rms(lc, model, meta, fitter):
         The metadata object.
     fitter : str
         The name of the fitter (for plot filename).
-
-    Notes
-    -----
-    History:
-
-    - December 29, 2021 Taylor Bell
-        Moved plotting code to a separate function.
-    - January 7-22, 2022 Megan Mansfield
-        Adding ability to do a single shared fit across all channels
     """
     if not isinstance(fitter, str):
         raise ValueError(f'Expected type str for fitter, instead received a '
@@ -411,7 +399,7 @@ def plot_rms(lc, model, meta, fitter):
         else:
             ch_number = str(channel).zfill(len(str(lc.nchannel)))
             fname_tag = f'ch{ch_number}'
-        fname = (f'figs{os.sep}fig5301_{fname_tag}_allanplot_{fitter}'
+        fname = (f'figs{os.sep}fig5301_{fname_tag}_RMS_TimeAveraging_{fitter}'
                  + plots.figure_filetype)
         plt.savefig(meta.outputdir+fname, bbox_inches='tight', dpi=300)
         if not meta.hide_plots:
@@ -433,13 +421,6 @@ def plot_corner(samples, lc, meta, freenames, fitter):
         The metadata object.
     fitter : str
         The name of the fitter (for plot filename).
-
-    Notes
-    -----
-    History:
-
-    - December 29, 2021 Taylor Bell
-        Moved plotting code to a separate function.
     """
     ndim = len(freenames)+1  # One extra for the 1D histogram
 
@@ -506,13 +487,6 @@ def plot_chain(samples, lc, meta, freenames, fitter='emcee', burnin=False,
     nthin : int; optional
         If >1, the plot will use every nthin point to help speed up
         computation and reduce clutter on the plot. Defaults to 1.
-
-    Notes
-    -----
-    History:
-
-    - December 29, 2021 Taylor Bell
-        Moved plotting code to a separate function.
     """
     nsubplots = nrows*ncols
     nplots = int(np.ceil(len(freenames)/nsubplots))
@@ -599,13 +573,6 @@ def plot_trace(trace, model, lc, freenames, meta, fitter='nuts', compact=False,
         Plot multidimensional variables in a single plot. Defailts to False.
     **kwargs : dict
         Additional keyword arguments to pass to pm.traceplot.
-
-    Notes
-    -----
-    History:
-
-    - November 22, 2022 Taylor Bell
-        Initial version.
     """
 
     max_subplots = az_rcParams['plot.max_subplots'] // 2
@@ -649,13 +616,6 @@ def plot_res_distr(lc, model, meta, fitter):
         The metadata object.
     fitter : str
         The name of the fitter (for plot filename).
-
-    Notes
-    -----
-    History:
-
-    - February 18, 2022 Caroline Piaulet
-        Created function
     """
     if not isinstance(fitter, str):
         raise ValueError(f'Expected type str for fitter, instead received a '
@@ -714,15 +674,6 @@ def plot_GP_components(lc, model, meta, fitter, isTitle=True):
         The name of the fitter (for plot filename).
     isTitle : bool; optional
         Should figure have a title. Defaults to True.
-
-    Notes
-    -----
-    History:
-
-    - February 28, 2022 Eva-Maria Ahrer
-        Written function
-    - March 9, 2022 Eva-Maria Ahrer
-        Adapted with shared parameters
     """
     if not isinstance(fitter, str):
         raise ValueError(f'Expected type str for fitter, instead received a '
@@ -1038,3 +989,65 @@ def plot_starry_star(lc, model, meta, fitter):
         fig.savefig(meta.outputdir+fname, bbox_inches='tight', dpi=300)
         if not meta.hide_plots:
             plt.pause(0.2)
+
+
+def plot_harmonica_string(lc, model, meta, fitter, isTitle=True):
+    """Plot the Harmonica string (Figs 5309)
+
+    Parameters
+    ----------
+    lc : eureka.S5_lightcurve_fitting.lightcurve.LightCurve
+        The lightcurve data object.
+    model : eureka.S5_lightcurve_fitting.models.CompositeModel
+        The fitted composite model.
+    meta : eureka.lib.readECF.MetaClass
+        The metadata object.
+    fitter : str
+        The name of the fitter (for plot filename).
+    isTitle : bool; optional
+        Should figure have a title. Defaults to True.
+    """
+    for c in range(lc.nchannel_fitted):
+        channel = lc.fitted_channels[c]
+        if lc.nchannel_fitted > 1:
+            chan = channel
+        else:
+            chan = 0
+
+        # Initialize PlanetParams object
+        pl_params = PlanetParams(model, 0, chan)
+
+        # Make the transit model
+        ht = HarmonicaTransit()
+        ht.set_planet_transmission_string(pl_params.ab)
+
+        # Compute the transmission string
+        theta = np.linspace(-np.pi, np.pi, 1000)
+        string = ht.get_planet_transmission_string(theta)
+
+        fig = plt.figure(5309, figsize=(8, 6))
+        plt.clf()
+        ax = fig.gca()
+        ax.set_aspect("equal", "datalim")
+        if isTitle:
+            ax.set_title(f'{meta.eventlabel} - Channel {channel} - '
+                         f'{fitter}')
+        plt.plot(string*np.cos(theta), string*np.sin(theta),
+                 c='C0', lw=2.5, label="Transmission string")
+        plt.plot(pl_params.ab[0] * np.cos(theta),
+                 pl_params.ab[0] * np.sin(theta),
+                 c='0.7', ls="--", label="Reference circle")
+        plt.xlabel("Planet x / Stellar Radius", fontsize=12)
+        plt.ylabel("Planet y / Stellar Radius", fontsize=12)
+
+        if lc.white:
+            fname_tag = 'white'
+        else:
+            ch_number = str(channel).zfill(len(str(lc.nchannel)))
+            fname_tag = f'ch{ch_number}'
+        fname = (f'figs{os.sep}fig5309_{fname_tag}_harmonica_string_{fitter}'
+                 + plots.figure_filetype)
+        fig.savefig(meta.outputdir+fname, bbox_inches='tight', dpi=300)
+        if not meta.hide_plots:
+            plt.pause(0.2)
+
