@@ -124,15 +124,21 @@ def fitlc(eventlabel, ecf_path=None, s4_meta=None, input_meta=None):
                     # Search
                     path = glob(path+os.sep+f'**{os.sep}*LCData.h5',
                                 recursive=True)
+
+                    # Get lightcurve file with latest modified time
+                    path = np.unique([os.sep.join(fname.split(os.sep))
+                                      for fname in path])
+
                     if len(path) == 0:
                         raise AssertionError(
                             'Unable to find any LCData save files at '
                             f'{path}')
                     elif len(path) > 1:
-                        print(f'WARNING: Found {len(path)} LCData save '
-                              f'files... Using {path[0]}')
-                    # Use the first file found
-                    path = path[0]
+                        print(f'WARNING: Found {len(path)} LCData save files'
+                              f'... Using {max(path, key=os.path.getmtime)}')
+
+                    # Use the latest file found
+                    path = max(path, key=os.path.getmtime)
                     lc_hold = xrio.readXR(path)
                     meta.wave_low = np.append(meta.wave_low,
                                               lc_hold.wave_low.values)
@@ -559,6 +565,7 @@ def fit_channel(meta, time, flux, chan, flux_err, eventlabel, params,
         BatmanTransitModel = m.BatmanTransitModel
         BatmanEclipseModel = m.BatmanEclipseModel
         CatwomanTransitModel = m.CatwomanTransitModel
+        HarmonicaTransitModel = m.HarmonicaTransitModel
         PoetTransitModel = m.PoetTransitModel
         PoetEclipseModel = m.PoetEclipseModel
         PoetPCModel = m.PoetPCModel
@@ -685,6 +692,25 @@ def fit_channel(meta, time, flux, chan, flux_err, eventlabel, params,
                                          num_planets=meta.num_planets,
                                          fac=meta.catwoman_fac,
                                          max_err=meta.catwoman_max_err)
+        modellist.append(t_transit)
+    if 'harmonica_tr' in meta.run_myfuncs:
+        t_transit = HarmonicaTransitModel(parameters=params,
+                                          fmt='r--', log=log, time=time,
+                                          time_units=time_units,
+                                          freenames=freenames,
+                                          longparamlist=lc_model.longparamlist,
+                                          nchannel=chanrng,
+                                          nchannel_fitted=nchannel_fitted,
+                                          fitted_channels=fitted_channels,
+                                          paramtitles=paramtitles,
+                                          ld_from_S4=meta.use_generate_ld,
+                                          ld_from_file=meta.ld_file,
+                                          ld_coeffs=ldcoeffs,
+                                          recenter_ld_prior=meta.recenter_ld_prior,  # noqa: E501
+                                          compute_ltt=meta.compute_ltt,
+                                          multwhite=lc_model.multwhite,
+                                          nints=lc_model.nints,
+                                          num_planets=meta.num_planets)
         modellist.append(t_transit)
     if 'fleck_tr' in meta.run_myfuncs:
         t_transit = m.FleckTransitModel(parameters=params,
@@ -917,7 +943,7 @@ def fit_channel(meta, time, flux, chan, flux_err, eventlabel, params,
         t_GP = GPModel(meta.kernel_class, meta.kernel_inputs, lc_model,
                        parameters=params, fmt='r--', log=log,
                        time=time, time_units=time_units,
-                       gp_code=meta.GP_package,
+                       gp_code_name=meta.GP_package,
                        useHODLR=meta.useHODLR,
                        freenames=freenames,
                        longparamlist=lc_model.longparamlist,
