@@ -25,6 +25,7 @@ if success:
     from . import S2_calibrations
 from . import S3_data_reduction
 from . import S4_generate_lightcurves
+from . import S4cal_StellarSpectra
 from . import S5_lightcurve_fitting
 from . import S6_planet_spectra
 
@@ -32,7 +33,8 @@ PACAKGEDIR = os.path.abspath(os.path.dirname(__file__))
 
 __all__ = ["lib", "S1_detector_processing", "S2_calibrations",
            "S3_data_reduction", "S4_generate_lightcurves",
-           "S5_lightcurve_fitting", "S6_planet_spectra"]
+           "S4cal_StellarSpectra", "S5_lightcurve_fitting",
+           "S6_planet_spectra"]
 
 # Make sure the required plotting setup is done even if the user doesn't
 # manually run the function
